Converted Climber for 3D paramater space(hardcoded) and parametric evaluation function. Created main in MyTools that uses the Climber statically to test the weight space

+++ src/student_player/mytools/StateNode.java
@@ -6,7 +6,7 @@ import hus.HusBoardState;
 import hus.HusMove;
 
 public class StateNode extends Node<HusBoardState> implements Cloneable{
-	private float evaluation;
+	private Double evaluation;
 	private boolean evaluated = false;
 	private int depth = -1;
 	private boolean myturn;
@@ -49,12 +49,12 @@ public class StateNode extends Node<HusBoardState> implements Cloneable{
 	}
 
 
-	public float getEvaluation() {
+	public Double getEvaluation() {
 		return evaluation;
 	}
 
 
-	public void setEvaluation(float evaluation) {
+	public void setEvaluation(Double evaluation) {
 		this.evaluation = evaluation;
 	}
